+++ src/main/java/com/Sydorenko/service/UserServiseImpl.java
@@ -27,7 +27,7 @@ private UserRepository userRepository;
     }
 
     @Override
-    public User findById ( int id ) {
+    public User findById ( long id ) {
         return userRepository.findById(id);
     }
